Buzzer added, remote control working properly, better performance compared to competition version, final version before semester began

+++ Devices/Devices.h/Motor_Init.h
@@ -21,7 +21,7 @@ typedef struct
 	 int16_t Actual_Current;
 	 int8_t Temperature;
 	 
-	 int16_t Target_Angle;
+	 int32_t Target_Angle;
 	 int16_t Target_Speed;
 	 int32_t Total_Angle;
 	 int16_t Turn_Count;
